requirements.txt
python-dotenv==0.21.0
boto3==1.26.16
google-cloud-storage==2.6.0